storybook thinks exports are stories, move constants to cont file
