+++ packages/webapp/src/stories/Modal/CropPlansModal.stories.jsx
@@ -12,10 +12,10 @@
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
  */
-import React from 'react';
 import CropPlansModal from '../../components/Modals/CropModals/CropPlansModal';
 import { componentDecorators } from '../Pages/config/Decorators';
 import { chromaticSmallScreen } from '../Pages/config/chromatic';
+import { commonArgs, planBaseContents } from './constants';
 
 export default {
   title: 'Components/Modals/CropPlansModal',
@@ -24,34 +24,6 @@ export default {
 };
 
 const Template = (args) => <CropPlansModal {...args} />;
-
-export const commonArgs = {
-  history: {
-    location: { pathname: '/crop/variety_id/management' },
-  },
-  match: {
-    params: {
-      variety_id: 'variety_id',
-    },
-  },
-  variety: {
-    crop_translation_key: 'Blueberry',
-    crop_variety_name: 'Nantes',
-    supplier: 'Buckerfields',
-    crop_variety_photo_url:
-      'https://litefarm.nyc3.cdn.digitaloceanspaces.com/default_crop/v2/blueberry.webp',
-  },
-  dismissModal: () => console.log('dismiss'),
-};
-
-export const planBaseContents = {
-  managementPlanName: 'Management Plan',
-  locationName: 'Field 1',
-  notes: 'Row 1',
-  numberOfPendingTask: 0,
-  status: 'active',
-  management_plan_group_id: '97643f51-6105-4462-aa21-a5048117017f',
-};
 
 export const ModalWithTwoCard = Template.bind({});
 ModalWithTwoCard.args = {
